Extract argparse definitions into a separate args.py module
Move the options() function from cli.py into a new args.py file,
reducing cli.py size and keeping argument definitions in one place.
